Move isDERSignature to TransactionSignature, and moved some tests around
