+++ src/Bitcoin/Signature/Signature.php
@@ -33,6 +33,7 @@ class Signature extends Serializable implements SignatureInterface
         $this
             ->setR($r)
             ->setS($s)
+
             ->setSighashType($sighashType);
     }
 
@@ -100,67 +101,5 @@ class Signature extends Serializable implements SignatureInterface
         $serializer = SignatureFactory::getSerializer();
         $buffer = $serializer->serialize($this);
         return $buffer;
-    }
-
-    /**
-     * @param \BitWasp\Buffertools\Buffer $sig
-     * @return bool
-     * @throws SignatureNotCanonical
-     */
-    public static function isDERSignature(Buffer $sig)
-    {
-        $checkVal = function ($fieldName, $start, $length, $binaryString) {
-            if ($length == 0) {
-                throw new SignatureNotCanonical('Signature ' . $fieldName . ' length is zero');
-            }
-            $typePrefix = ord(substr($binaryString, $start - 2, 1));
-            if ($typePrefix !== 0x02) {
-                throw new SignatureNotCanonical('Signature ' . $fieldName . ' value type mismatch');
-            }
-            $val = substr($binaryString, $start, $length);
-            $vAnd = $val[0] & pack("H*", '80');
-            if (ord($vAnd) === 128) {
-                throw new SignatureNotCanonical('Signature ' . $fieldName . ' value is negative');
-            }
-            if ($length > 1 && ord($val[0]) == 0x00 && !ord(($val[1] & pack('H*', '80')))) {
-                throw new SignatureNotCanonical('Signature ' . $fieldName . ' value excessively padded');
-            }
-        };
-
-        $bin = $sig->getBinary();
-        $size = $sig->getSize();
-        if ($size < 9) {
-            throw new SignatureNotCanonical('Signature too short');
-        }
-
-        if ($size > 73) {
-            throw new SignatureNotCanonical('Signature too long');
-        }
-
-        if (ord($bin[0]) !== 0x30) {
-            throw new SignatureNotCanonical('Signature has wrong type');
-        }
-
-        if (ord($bin[1]) !== $size - 3) {
-            throw new SignatureNotCanonical('Signature has wrong length marker');
-        }
-
-        $lenR = ord($bin[3]);
-        $startR = 4;
-        if (5 + $lenR >= $size) {
-            throw new SignatureNotCanonical('Signature S length misplaced');
-        }
-
-        $lenS = ord($bin[5 + $lenR]);
-        $startS = 4 + $lenR + 2;
-        if (($lenR + $lenS + 7) !== $size) {
-            throw new SignatureNotCanonical('Signature R+S length mismatch');
-        }
-
-        $checkVal('R', $startR, $lenR, $bin);
-        $checkVal('S', $startS, $lenS, $bin);
-
-        return true;
-
     }
 }
